Move 134 previously nightly-only tests that are actually "fast enough" to be included in the normal CI runs. Move 66 actually extra slow tests to an "extended_tests" module.
