+++ core/service/src/client/tests/centralized/mod.rs
@@ -1,10 +1,10 @@
 mod crs_gen_tests;
 mod custodian_backup_tests;
 mod custodian_context_tests;
+#[cfg(feature = "slow_tests")]
+mod extended_tests;
 mod key_gen_tests;
 mod misc_tests;
-#[cfg(feature = "slow_tests")]
-mod nightly_tests;
 mod public_decryption_tests;
 mod restore_from_backup_tests;
 mod user_decryption_tests;
