+++ core/service/src/client/tests/threshold/mod.rs
@@ -2,13 +2,14 @@ mod common;
 mod crs_gen_tests;
 mod custodian_backup_tests;
 mod custodian_context_tests;
+// These are slow tests that we run in the nightly scheduled run. Some are very slow.
+#[cfg(feature = "slow_tests")]
+mod extended_tests;
 mod key_gen_tests;
 mod misc_tests;
 mod mpc_context_tests;
 #[cfg(feature = "slow_tests")]
 mod mpc_epoch_tests;
-#[cfg(feature = "slow_tests")]
-mod nightly_tests;
 mod public_decryption_tests;
 mod restore_from_backup_tests;
 mod user_decryption_tests;
