+++ core/service/src/client/tests/threshold/restore_from_backup_tests.rs
@@ -42,7 +42,7 @@ use tokio::task::JoinSet;
 ///
 /// **Requires:** `insecure` feature flag
 #[tokio::test]
-async fn nightly_test_insecure_threshold_dkg_backup() -> Result<()> {
+async fn test_insecure_threshold_dkg_backup() -> Result<()> {
     // Setup using builder pattern with backup vault
     let env = ThresholdTestEnv::builder()
         .with_test_name("threshold_dkg_backup")
@@ -252,7 +252,7 @@ async fn nightly_test_insecure_threshold_dkg_backup() -> Result<()> {
 ///
 /// **Requires:** `insecure` feature flag
 #[tokio::test]
-async fn nightly_test_insecure_threshold_autobackup_after_deletion() -> Result<()> {
+async fn test_insecure_threshold_autobackup_after_deletion() -> Result<()> {
     // Setup using builder pattern with backup vault
     let env = ThresholdTestEnv::builder()
         .with_test_name("threshold_autobackup")
